Use the Location header if it available
Faust service PR  https://github.com/grame-cncm/faustservice/pull/10 implements a
scheme whereby the location of a generated artifact is returned in the Location header.

This allows us, in the IDE, to be independent of the artifact naming implemented by
the faust service. This PR exploits that feature, while still retaining the hard-coded
`binary.zip` artifact name for when the Location header is not present.